chore: organize format functions
